refactor: move copy_ensuring_int_transaction_type to util
