+++ cfx_account/_utils/transactions/base.py
@@ -2,9 +2,6 @@ from abc import ABC, abstractmethod
 from typing import Any, Dict, Tuple
 from typing_extensions import Self
 from cfx_utils.types import TxParam
-from toolz import (
-    assoc,
-)
 
 class TransactionImplementation(ABC):
     """
@@ -43,24 +40,3 @@ class TransactionImplementation(ABC):
     @abstractmethod
     def encode(self) -> bytes:
         pass
-
-# returns a copy of the transaction dict with the 'type' field converted to int
-def copy_ensuring_int_transaction_type(transaction_dict: TxParam) -> Dict[str, Any]:
-    if "type" not in transaction_dict:
-        if "gasPrice" in transaction_dict:
-            if "accessList" in transaction_dict:
-                # access list txn - type 1
-                return assoc(transaction_dict, "type", 1)
-            else:
-                return assoc(transaction_dict, "type", 0)
-        # elif any(
-        #     type_2_arg in transaction_dict
-        #     for type_2_arg in ("maxFeePerGas", "maxPriorityFeePerGas")
-        # ):
-        else:
-            return assoc(transaction_dict, "type", 2)
-    cpy = transaction_dict.copy()
-    if isinstance(cpy["type"], str):
-        cpy["type"] = int(cpy["type"], 16)
-    return cpy
-
